python-mysql: rename to python-mysqlclient & bump to version 1.4.4
The package on PyPi is named `mysqlclient`.
This should have been named `python-mysqlclient` from the start.
There is a `mysql` package on PyPi already but that's a different
code/package.

Doing this should avoid any future confusion.
There is no good time to do this rename; at least 19.07 has been branched
already and this can go into the next release [in a year or so].

Signed-off-by: Alexandru Ardelean <ardeleanalex@gmail.com>